// src/app/core/models/auth.models.ts
export interface User {
  id: string;
  name: string;
  email: string;
  role?: 'ADMIN' | 'CUSTOMER' | string;
  avatar?: string | null;
  phone?: string | null;
}

export interface GoogleLoginRequest {
  credential: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface LoginResponse {
  accessToken: string;
  refreshToken?: string;
  user: User;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}